FIxed crashes with really ugly files
We can work now even with files that are totaly ugly (like relative
files, with wrongly formated lines and strange consecutive sylabels that
use a wrong format.

They should be killed with fire however, so we do not support writing
relative files.

TODO: Strip the relative tag in the outpit file (they are not relative!)